Move inquiry respond related functions to service module
Refactor the inquiry API endpoint and move related functions for responding to an inquiry into the inquiry service module. This allows for better reuse and unit testing. The design of inquiry API and DB models are not refactored. The functions in the inquiry service module will expect the inquiry API model.